requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.13.3
alembic-postgresql-enum==1.5.0
alembic_utils==0.8.6
annotated-types==0.7.0
anyio==4.8.0
apispec==6.8.1
argcomplete==3.5.3
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.1.0
babel==2.17.0
bcrypt==4.0.1
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.2
debugpy==1.8.12
Deprecated==1.2.18
dill==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.7
fastapi-cache2==0.2.2
fastapi-cli==0.0.7
fastapi-filter==2.0.1
fastapi-pagination==0.13.2
Flask==2.2.5
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
flupy==1.2.1
frozenlist==1.5.0
fsspec==2025.2.0
GeoAlchemy2==0.17.0
geopandas==1.0.1
google-re2==1.1.20240702
googleapis-common-protos==1.69.0
greenlet==3.1.1
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lark==1.2.2
lazy-object-proxy==1.10.0
limits==4.0.1
linkify-it-py==2.0.3
lockfile==0.12.2
loguru==0.7.2
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==1.4.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.6.0
multidict==6.1.0
numpy==2.2.2
ordered-set==4.1.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parse==1.20.2
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.5.0
prison==0.2.1
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg-pool==3.2.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.19.1
PyJWT==2.10.1
pyogrio==0.10.0
pyproj==3.7.1
pyserial==3.5
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
redis==4.6.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.0
rich-toolkit==0.13.2
rpds-py==0.23.1
rsa==4.9
ruff==0.9.3
setproctitle==1.3.5
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlalchemy-filters==0.13.0
sqlalchemy-filterset==2.3.0
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.24
sqlparse==0.5.3
starlette==0.45.3
suds==1.2.0
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
time-machine==2.16.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
ujson==5.10.0
universal_pathlib==0.2.6
urllib3==2.3.0
uvicorn==0.31.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
yarl==1.18.3
yattag==1.16.1
zipp==3.21.0
SQLAlchemy-serializer